@@ -2,26 +2,40 @@ package com.mdff.app.activity;
 
 import android.Manifest;
 import android.app.Activity;
+import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Typeface;
+import android.media.AudioManager;
+import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
+import android.os.Handler;
+import androidx.core.app.ActivityCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
+import android.view.WindowManager;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.FrameLayout;
+import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
+import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.android.volley.DefaultRetryPolicy;
+import com.android.volley.Request;
+import com.android.volley.Response;
+import com.android.volley.RetryPolicy;
+import com.android.volley.TimeoutError;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.JsonObjectRequest;
 import com.mdff.app.R;
 import com.mdff.app.app_interface.VolleyCallback;
 import com.mdff.app.controller.AlphaApplication;
@@ -32,18 +46,13 @@ import com.mdff.app.utility.AppUtil;
 import com.mdff.app.utility.ConnectivityReceiver;
 import com.mdff.app.utility.Constant;
 import com.mdff.app.utility.NetworkUtils;
-import com.android.volley.DefaultRetryPolicy;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.RetryPolicy;
-import com.android.volley.TimeoutError;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
+import com.mdff.app.utility.Utilities;
 import com.squareup.picasso.Picasso;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.io.IOException;
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
@@ -52,7 +61,7 @@ import java.util.Map;
  * Created by Deepika.Mishra on 5/29/2018.
  */
 
-public class MessageDetails extends AppCompatActivity implements VolleyCallback {
+public class MessageDetails extends AppCompatActivity implements VolleyCallback, SeekBar.OnSeekBarChangeListener {
     Activity activity;
     ImageView iv_message, iv_play;
     WebView wv_description;
@@ -63,17 +72,39 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
     MessageItems data;
     private String comeFrom = "";
     int connectStatus;
-//    iframe
+    //    iframe
     private MyWebChromeClient mWebChromeClient = null;
     private View mCustomView;
     private RelativeLayout mContentView;
     private FrameLayout mCustomViewContainer;
     private WebChromeClient.CustomViewCallback mCustomViewCallback;
 
+
+    private ImageButton forwardbtn, backwardbtn, pausebtn, playbtn;
+    private MediaPlayer mPlayer;
+    private TextView songName, startTime, songTime;
+    private SeekBar songPrgs;
+    private static int oTime = 0, sTime = 0, eTime = 0, fTime = 5000, bTime = 5000;
+    private Handler hdlr = new Handler();
+    private RelativeLayout rl_audio, imv_layout;
+    private ImageView imgLogo;
+    private ProgressDialog progressDialog;
+    private int seekForwardTime = 5000; // 5000 milliseconds
+    private int seekBackwardTime = 5000; // 5000 milliseconds
+
+
+    Utilities utils;
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.fragment_messsage_detail);
+        getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        mPlayer = new MediaPlayer();
+        utils = new Utilities();
+
+
         Intent i = getIntent();
         data = (MessageItems) i.getSerializableExtra("messagedetails");
         comeFrom = i.getStringExtra("comeFrom");
@@ -83,9 +114,187 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
         getMessageStatusUpdated(data);
         initializeIds();
 
+        rl_audio = findViewById(R.id.rl_audio);
+        imv_layout = findViewById(R.id.imv_layout);
+
+
+        backwardbtn = (ImageButton) findViewById(R.id.btnBackward);
+        forwardbtn = (ImageButton) findViewById(R.id.btnForward);
+        playbtn = (ImageButton) findViewById(R.id.btnPlay);
+        pausebtn = (ImageButton) findViewById(R.id.btnPause);
+
+        startTime = (TextView) findViewById(R.id.txtStartTime);
+        songTime = (TextView) findViewById(R.id.txtSongTime);
+
+        songPrgs = (SeekBar) findViewById(R.id.sBar);
+        songPrgs.setOnSeekBarChangeListener(this);
+        songPrgs.setClickable(true);
+
+
+        pausebtn.setEnabled(false);
+
+        if (data.getType().equalsIgnoreCase("audio")) {
+            rl_audio.setVisibility(View.VISIBLE);
+
+            progressDialog = new ProgressDialog(this);
+            progressDialog.setMessage("Loading...");
+            progressDialog.setCancelable(false);
+            String url = data.getAttachment_url(); // your URL here
+            progressDialog.show();
+
+            mPlayer.setAudioStreamType(AudioManager.STREAM_MUSIC);
+            try {
+                mPlayer.setDataSource(url);
+                mPlayer.prepareAsync();
+                mPlayer.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
+                    @Override
+                    public void onPrepared(MediaPlayer mp) {
+                        if (mp.equals(mPlayer)) {
+                            updateProgressBar();
+                        }
+                        if (progressDialog != null && progressDialog.isShowing())
+                            progressDialog.dismiss();
+
+                    }
+                });
+
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+
+        mPlayer.setOnBufferingUpdateListener(new MediaPlayer.OnBufferingUpdateListener() {
+            @Override
+            public void onBufferingUpdate(MediaPlayer mp, int percent) {
+               /* final Toast toast = Toast.makeText(MessageDetails.this, "Buffering...", Toast.LENGTH_SHORT);
+                toast.show();*/
+                Handler handler = new Handler();
+                handler.postDelayed(new Runnable() {
+                    @Override
+                    public void run() {
+
+                    }
+                }, 500);
+            }
+        });
+
+
+        if (data.getType().equals("audio")) {
+
+            String imgUri = data.getThumbnail_url();
+            if (imgUri != null && !imgUri.equals("")) {
+                Picasso.get().load(imgUri).into(imgLogo);
+            }
+
+
+          /*  rl_audio.setVisibility(View.VISIBLE);
+            imv_layout.setVisibility(View.GONE);*/
+        } else {
+           /* rl_audio.setVisibility(View.GONE);
+            imv_layout.setVisibility(View.VISIBLE);*/
+        }
+
+
+        playbtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mPlayer.start();
+                sTime = mPlayer.getCurrentPosition();
+                if (oTime == 0) {
+                    songPrgs.setMax(eTime);
+                    oTime = 1;
+                }
+
+                songPrgs.setProgress(sTime);
+                updateProgressBar();
+                pausebtn.setEnabled(true);
+                playbtn.setEnabled(false);
+            }
+        });
+
+        pausebtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mPlayer.pause();
+                pausebtn.setEnabled(false);
+                playbtn.setEnabled(true);
+            }
+        });
+
+        forwardbtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                // get current song position
+                int currentPosition = mPlayer.getCurrentPosition();
+                // check if seekForward time is lesser than song duration
+                if (currentPosition + seekForwardTime <= mPlayer.getDuration()) {
+                    // forward song
+                    mPlayer.seekTo(currentPosition + seekForwardTime);
+                } else {
+                    // forward to end position
+                    mPlayer.seekTo(mPlayer.getDuration());
+                }
+
+                if (!playbtn.isEnabled()) {
+                    playbtn.setEnabled(true);
+                }
+            }
+        });
+
+        backwardbtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // get current song position
+                int currentPosition = mPlayer.getCurrentPosition();
+                // check if seekBackward time is greater than 0 sec
+                if (currentPosition - seekBackwardTime >= 0) {
+                    // forward song
+                    mPlayer.seekTo(currentPosition - seekBackwardTime);
+                } else {
+                    // backward to starting position
+                    mPlayer.seekTo(0);
+                }
+                if (!playbtn.isEnabled()) {
+                    playbtn.setEnabled(true);
+                }
+            }
+        });
+
+
     }
 
+    public void updateProgressBar() {
+        hdlr.postDelayed(mUpdateTimeTask, 100);
+    }
+
+    private Runnable mUpdateTimeTask = new Runnable() {
+        public void run() {
+            long totalDuration = mPlayer.getDuration();
+            long currentDuration = mPlayer.getCurrentPosition();
+            // Displaying Total Duration time
+
+            // Displaying time completed playing
+
+            // Displaying Total Duration time
+            songTime.setText("" + utils.milliSecondsToTimer(totalDuration));
+            // Displaying time completed playing
+            startTime.setText("" + utils.milliSecondsToTimer(currentDuration));
+
+            // Updating progress bar
+            int progress = (int) (utils.getProgressPercentage(currentDuration, totalDuration));
+            //Log.d("Progress", ""+progress);
+            songPrgs.setProgress(progress);
+
+            // Running this thread after 100 milliseconds
+            hdlr.postDelayed(this, 100);
+        }
+    };
+
+
     private void initializeIds() {
+        imgLogo = (ImageView) findViewById(R.id.imgLogo);
         backLayout = (LinearLayout) findViewById(R.id.backLayout);
         iv_message = (ImageView) findViewById(R.id.iv_message);
         iv_play = (ImageView) findViewById(R.id.iv_play);
@@ -125,23 +334,28 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
 
 
             if (data.getType().equalsIgnoreCase("document")) {
-                Picasso.with(this).load(data.getThumbnail_url())//download URL
+                Picasso.get().load(data.getThumbnail_url())//download URL
                         .into(iv_message);//imageview
                 iv_message.setVisibility(View.VISIBLE);
             } else if (data.getType().equalsIgnoreCase("image")) {
-                Picasso.with(this).load(data.getAttachment_url())//download URL
+                Picasso.get().load(data.getAttachment_url())//download URL
                         .into(iv_message);//imageview
                 iv_message.setVisibility(View.VISIBLE);
 
 
             } else if (data.getType().equalsIgnoreCase("video")) {
-                Picasso.with(this).load(data.getThumbnail_url())//download URL
+                Picasso.get().load(data.getThumbnail_url())//download URL
                         .into(iv_message);//imageview
                 iv_message.setVisibility(View.VISIBLE);
                 iv_play.setVisibility(View.VISIBLE);
 
+            } else if (data.getType().equalsIgnoreCase("audio")) {
+                rl_audio.setVisibility(View.VISIBLE);
+                Picasso.get().load(data.getThumbnail_url()).into(imgLogo);//imageview
+              //  rl_audio.setVisibility(View.VISIBLE);
+
             } else {
-                iv_message.setVisibility(View.GONE);
+             //   iv_message.setVisibility(View.GONE);
 
             }
 
@@ -166,8 +380,7 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
             @Override
             public void onClick(View view) {
 
-                if (data.getType().equalsIgnoreCase("document"))
-                {
+                if (data.getType().equalsIgnoreCase("document")) {
                     try {
                         Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(data.getAttachment_url()));
                         startActivity(browserIntent);
@@ -203,6 +416,29 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
             alertMessage.show(getFragmentManager(), "");
 
         }
+    }
+
+
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+
+    }
+
+    @Override
+    public void onStartTrackingTouch(SeekBar seekBar) {
+        hdlr.removeCallbacks(mUpdateTimeTask);
+    }
+
+    @Override
+    public void onStopTrackingTouch(SeekBar seekBar) {
+        hdlr.removeCallbacks(mUpdateTimeTask);
+        int totalDuration = mPlayer.getDuration();
+        int currentPosition = utils.progressToTimer(seekBar.getProgress(), totalDuration);
+        // forward or backward to certain seconds
+        mPlayer.seekTo(currentPosition);
+
+        // update timer progress again
+        updateProgressBar();
     }
 
 
@@ -286,7 +522,7 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
                     phoneCall(toUserMobileNo);
 
                 } else {
-//            Toast.makeText(this, "Required permissions are not granted", Toast.LENGTH_LONG).show();
+
                 }
                 break;
 
@@ -401,9 +637,7 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
 
             }
 
-        } else
-
-        {
+        } else {
             AlertMessage alertMessage = AlertMessage.newInstance(
                     getString(R.string.noInternet), getString(R.string.ok), getString(R.string.alert));
             alertMessage.show(getFragmentManager(), "");
@@ -423,7 +657,7 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
                 callback.onCustomViewHidden();
                 return;
             }
-            try{
+            try {
                 mContentView = (RelativeLayout) findViewById(R.id.detail);
 //            mContentView.setBackgroundColor(getResources().getColor(R.color.black));
                 mContentView.setVisibility(View.GONE);
@@ -440,19 +674,21 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
                 mCustomViewContainer.setVisibility(View.VISIBLE);
                 setContentView(mCustomViewContainer);
 
+            } catch (Exception e) {
             }
-            catch (Exception e){}
         }
+
         public void setActivityBackgroundColor(int color) {
             View view = activity.getWindow().getDecorView();
             view.setBackgroundColor(color);
         }
+
         @Override
         public void onHideCustomView() {
             if (mCustomView == null) {
                 return;
             } else {
-                try{
+                try {
                     // Hide the custom view.
 
                     mCustomView.setVisibility(View.GONE);
@@ -461,15 +697,15 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
                     mCustomView = null;
 
                     mCustomViewContainer.setVisibility(View.GONE);
-                    mCustomViewContainer=null;
+                    mCustomViewContainer = null;
                     mCustomViewCallback.onCustomViewHidden();
                     // Show the content view.
 //                mContentView.setBackgroundColor(getResources().getColor(R.color.white));
                     mContentView.setVisibility(View.VISIBLE);
                     setContentView(mContentView);
 //                    setActivityBackgroundColor(getResources().getColor(R.color.white));
+                } catch (Exception e) {
                 }
-                catch (Exception e){}
             }
         }
     }
@@ -479,20 +715,33 @@ public class MessageDetails extends AppCompatActivity implements VolleyCallback 
         if (mCustomViewContainer != null)
             try {
                 mWebChromeClient.onHideCustomView();
+            } catch (Exception e) {
             }
-            catch (Exception e){}
         else if (wv_description.canGoBack())
-            try{
+            try {
                 wv_description.goBack();
+            } catch (Exception e) {
             }
-            catch (Exception e){}
         else
-            try{
+            try {
                 Intent intent = new Intent(activity, Home.class);
                 intent.putExtra("comeFrom", "messageDetail");
                 startActivity(intent);
                 finish();
+            } catch (Exception e) {
             }
-            catch (Exception e){}
+    }
+
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        try {
+            mPlayer.stop();
+            mPlayer.release();
+            hdlr.removeCallbacks(mUpdateTimeTask);
+        } catch (Exception ex) {
+
+        }
     }
 }
