@@ -4,9 +4,9 @@ import android.app.Activity;
 import android.app.DialogFragment;
 import android.content.Intent;
 import android.graphics.Typeface;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -417,6 +417,7 @@ public class ResourceSubList extends AppCompatActivity implements VolleyCallback
                                                         resource.setTopic_id(jsonObject1.getString("topic_id"));
                                                         resource.setImage(jsonObject1.getString("image"));
                                                         resource.setVideo_url(jsonObject1.getString("video_url"));
+                                                        resource.setAudio_url(jsonObject1.getString("audio_url"));
                                                         resource.setThumbnail_url(jsonObject1.getString("thumbnail_url"));
                                                         resource.setType(jsonObject1.getString("type"));
                                                         resource.setTitle(jsonObject1.getString("title"));
