@@ -4,11 +4,17 @@ import android.app.Activity;
 import android.app.DialogFragment;
 import android.app.ProgressDialog;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.Toast;
 
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
 import com.mdff.app.R;
 import com.mdff.app.controller.AlphaApplication;
 
@@ -243,4 +249,26 @@ public class Splash extends AppCompatActivity implements AlertMessage.NoticeDial
         System.exit(0);
 
     }
+
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        FirebaseInstanceId.getInstance().getInstanceId()
+                .addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<InstanceIdResult> task) {
+                        if (!task.isSuccessful()) {
+                            return;
+                        }
+                        // Get new Instance ID token
+                        String token = task.getResult().getToken();
+                        Log.d("Token",">>"+token);
+                        appUtil.setPrefrence("deviceToken",token);
+
+                    }
+                });
+    }
 }
+
+
