@@ -6,9 +6,9 @@ import android.content.Context;
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
@@ -527,6 +527,7 @@ public class Resources extends Fragment implements VolleyCallback, ResoucrceInte
                                                         resource.setTopic_id(jsonObject1.getString("topic_id"));
                                                         resource.setImage(jsonObject1.getString("image"));
                                                         resource.setVideo_url(jsonObject1.getString("video_url"));
+                                                        resource.setAudio_url(jsonObject1.getString("audio_url"));
                                                         resource.setThumbnail_url(jsonObject1.getString("thumbnail_url"));
                                                         resource.setType(jsonObject1.getString("type"));
                                                         resource.setTitle(jsonObject1.getString("title"));
