@@ -5,46 +5,47 @@ package com.mdff.app.utility;
  */
 
 public class Constant {
-    public static final String tokenizationKey="sandbox_x3kyw6cg_cj8c5nxgfr7j8h8z";
 
-    public static final String APP_ID="MjAxOTA2MjQ1OTA1MjQ=";
+    public static final String tokenizationKey = "sandbox_x3kyw6cg_cj8c5nxgfr7j8h8z";
+
+    public static final String APP_ID = "MjAxOTA2MjQ1OTA1MjQ=";
 
     public static final String API_DEV_KEY = "AIzaSyBRVuMNvddWal28t99gmIeHqHe0wU5jM30";
-   // public static final String STRIPE_API_PUB_KEY = "pk_test_bosv5Urd7sXLEd8xvJJYQtdX";
+
+
+    /*public static final String WEB_URL = "https://1qa.iamanalpha1.com/api/";   //QA AWS server
+    public static final String STRIPE_API_PUB_KEY = "pk_test_bosv5Urd7sXLEd8xvJJYQtdX";*/
+
+    public static final String WEB_URL = "https://iamanalpha1.com/api/";     // live AWS server
     public static final String STRIPE_API_PUB_KEY = "pk_live_KSlcTs97CPG9B8uIJCIqRgdd";
 
-    //public static final String WEB_URL="http://iaaa.srmtechsol.com/api/";   // SRM server
-    public static final String WEB_URL="https://iamanalpha1.com/api/";     // live AWS server
-  //  public static final String WEB_URL="https://1qa.iamanalpha1.com/api/";   //QA AWS server     http://54.85.241.222/
-    public static final String APP_KEY="appkey";
-    public static final String SIGN_UP="register";
-    public static final String LOGIN="login";
-    public static final String FEED="getfeeds";
-    public static final String RESOURCE="getTopics";
-    public static final String RESOURCE_CONTENT="getContent";
-    public static final String RESOURCE_SUB_TOPICS="getSubtopics";
-    public static final String MESSAGE="getMessages";
-    public static final String MESSAGEREAD="messageRead";
-    public static final String UNREADMESSAGE="unreadMessages";
-    public static final String RESET_PWD="resetpassword";
-    public static final String UPDATE_PWD="updatepassword";
-    public static final String CHECK_MAIL="checkemail";
-    public static final String GET_QA="getQA";
-    public static final String INSERT_Q="insertQA";
-    public static final String CHECK_USER_NAME="checkUsername";
-    public static final String PLAN="planLists";
-    public static final String CARD_ADD="cardAdd";
-    public static final String CARD_LIST="cardList";
-    public static final String CARD_DELETE="cardDelete";
-    public static final String CANCEL_SUBSCRIPTION="cancelSubscription";
-    public static final String GETPROFILE="getProfile";
-    public static final String UPDATEPROFILE="updateProfile";
-    public static final String GET_COMMENTS="getComments";
-    public static final String POST_COMMENTS="postComment";
-    public static final String LIKE_POST="likePost";
 
-
-
+    public static final String APP_KEY = "appkey";
+    public static final String SIGN_UP = "register";
+    public static final String LOGIN = "login";
+    public static final String FEED = "getfeeds";
+    public static final String RESOURCE = "getTopics";
+    public static final String RESOURCE_CONTENT = "getContent";
+    public static final String RESOURCE_SUB_TOPICS = "getSubtopics";
+    public static final String MESSAGE = "getMessages";
+    public static final String MESSAGEREAD = "messageRead";
+    public static final String UNREADMESSAGE = "unreadMessages";
+    public static final String RESET_PWD = "resetpassword";
+    public static final String UPDATE_PWD = "updatepassword";
+    public static final String CHECK_MAIL = "checkemail";
+    public static final String GET_QA = "getQA";
+    public static final String INSERT_Q = "insertQA";
+    public static final String CHECK_USER_NAME = "checkUsername";
+    public static final String PLAN = "planLists";
+    public static final String CARD_ADD = "cardAdd";
+    public static final String CARD_LIST = "cardList";
+    public static final String CARD_DELETE = "cardDelete";
+    public static final String CANCEL_SUBSCRIPTION = "cancelSubscription";
+    public static final String GETPROFILE = "getProfile";
+    public static final String UPDATEPROFILE = "updateProfile";
+    public static final String GET_COMMENTS = "getComments";
+    public static final String POST_COMMENTS = "postComment";
+    public static final String LIKE_POST = "likePost";
 
 
 }
